requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain>=0.1.16,<0.2.0
langchain-community>=0.0.34,<0.1.0
langchain-google-genai>=0.0.6
google-generativeai>=0.3.2
faiss-cpu==1.13.2
pydantic==2.6.4
tiktoken==0.4.0
